Allows servers to be a Hash or Array
Chef best practices recommend that node attributes are kept in hashes
because of the sometimes confusing way array attributes are handled.
For example, there is no natural way to override the values inside
a hash which is part of an array attribute. This makes it difficult
to break up the configuration of Redis instances between roles,
environments and recipes.

The complications are obvious from the fact that both the README and
the code of the cookbook makes note of special handling related to
array attributes.

This commit allows hashes to be used instead of arrays while
maintaining full backward compatibility.

- Adds `RedisioHelper.each_server`, which acts as an adapter that yields
  a server info `Mash` regardless of the underlying collection.

- Changes `servers` resource attribute type to `Enumerable`.

- Localizes `node.set` for `:servers` to the case when it is `nil`.

- Updates changelog.

- Updates README with server hash as the preferred syntax.